First day additions
First day of the game jam is now over. I started the jam by writing a GDD, where I planned the core mechanics and idea of the game. The GDD remains unfinished, for now. After some planning, I began with graphics and then focused on programming. Here is a list of all the things I added (that I can recall):

+ Added graphics for 1st boat, 1st island, docks, water, 1st tree and 3 grass variations
+ Added a minimap and some markers
+ Added some folders -> organized
+ Added movement for the boat, along with a boat health script that detects hard collisions
+ Added Camera follow script

I took the first day a bit more chill, but I'm still proud of the amount of work I did.

- 12.5, 01.39 (UTC +3)