Second day
Second day of the game jam is now over. On this day, I focused mostly on the Delivery User Interface and programming the core harvesting mechanics. I also changed the title of the game to "Resilient Waters". Here's a list of all the things I added and changed today (that I can recall):

+ Added graphics for 1st big rock, credit coin, delivery UI buttons, 1st package, 1st contraband package, 1st scanned package, and 1st scanned contraband package
+ Added Big Rock
+ Came up with names for all the islands
+ Added the base mechanics for the Delivery UI system
+ Added ObjectSpawner script that can randomize Island and BigRock positions
+ Added the base mechanics for the harvesting system

All in all, the second day was productive, though I was expecting to see a bit more development for the game play.

- 13.5, 02.15 (UTC +3)