Day 5
Day 5 is now over. I focused on creating the boat upgrading system and it is now almost complete. I also added a bunch of other cool stuff. Here's a list of all the things I added today (that I can recall):

* Added Boat upgrading system
* Added Home island with a boat shop and docks/game start menu
* Added night shift and day shift functionality
* Added graphics for Home island, one other island, _Nova's_ home, and vertical docks
* Added different level water guns
* Added Pass out functionality that warns _Nova_ of the time one hour before passing out. When _Nova_ passes out, text appears on screen, and the scene changes to HomeIsland
* COMPOSED, RECORDED & ADDED MENU MUSIC!!!

All in all, it was a very challenging but rewarding day of coding, drawing, and composing. I am most proud of the music. I was so tired that I decided to go lay down, but got stressed I couldn't finish the game on time, so I decided to grab my guitar and the first thing I played came to be the Main Menu music, after about 1 hour of playing around with different sounds in FL Studio. Very happy with the outcome and how fast I was able to come up with it. Looking forward to tomorrow, even though I probably won't be able to finish the game, yet.

* 16.5, 2.05 (UTC +3)