Final merge
The jam is now over. I didn't update anything after the 7th day, because I had no time and was super tired after every night :D.
But a finished project is a project I can be proud of. There are still many bugs to fix and other minor improvements to be added. I will most likely continue development, I have many new ideas for this game. But, we'll see.

- 20.5, 19.10 (UCT +3)